feat: add loom-dl.js script for downloading videos from loom.com chore(package.json): update main script to loom-dl.js and add loom-dl to bin
The new loom-dl.js script allows users to download videos from loom.com either individually or in bulk from a list. It supports various command-line options for specifying the video URL or list file, output filename or directory, download timeout, and filename prefix. The main script in package.json has been updated to point to this new script, and a new bin entry has been added to allow the script to be run as a command-line tool. This enhances the functionality of the package and makes it more user-friendly.